Extract wallet util functions to utils/walletUtils.ts
